+++ packages/app/src/render.tsx
@@ -7,9 +7,9 @@ import { DisableShadowRootContext } from '@masknet/theme'
 import { MainUI } from './MainUI.js'
 import { useTheme } from './hooks/useTheme.js'
 import { initWallet } from '@masknet/web3-providers'
-import { createSharedContext } from './helpers/createSharedContext.js'
+import { WalletIO } from './helpers/createSharedContext.js'
 
-initWallet(createSharedContext())
+initWallet(WalletIO)
 const root = document.createElement('main')
 document.body.appendChild(root)
