+++ packages/mask/src/extension/dashboard/load-dashboard.tsx
@@ -3,29 +3,17 @@ import { startPluginDashboard } from '@masknet/plugin-infra/dashboard'
 import { DashboardRoutes } from '@masknet/shared-base'
 import { createSharedContext, createPluginHost } from '../../../shared/plugin-infra/host.js'
 import { createNormalReactRoot } from '../../utils/createNormalReactRoot.js'
-import { NextSharedUIContext, RestPartOfPluginUIContextShared } from '../../utils/plugin-context-shared-ui.js'
+import { allPersonas } from '../../../shared-ui/initUIContext.js'
 import Services from '#services'
 import { Modals } from '@masknet/shared'
-import { __setUIContext__ } from '@masknet/plugin-infra/dom'
 
-__setUIContext__({
-    allPersonas: NextSharedUIContext.allPersonas,
-    currentPersona: NextSharedUIContext.currentPersonaIdentifier,
-    queryPersonaAvatar: Services.Identity.getPersonaAvatar,
-    querySocialIdentity: Services.Identity.querySocialIdentity,
-    fetchJSON: Services.Helper.fetchJSON,
-    queryPersonaByProfile: Services.Identity.queryPersonaByProfile,
-    openDashboard: Services.Helper.openDashboard,
-    openPopupWindow: Services.Helper.openPopupWindow,
-    signWithPersona: (a, b, c, d) => Services.Identity.signWithPersona(a, b, c, location.origin, d),
-})
 startPluginDashboard(
     createPluginHost(
         undefined,
         (id, def, signal) => ({
             ...createSharedContext(id, signal),
-            ...RestPartOfPluginUIContextShared,
-            allPersonas: NextSharedUIContext.allPersonas,
+            hasPaymentPassword: Services.Wallet.hasPassword,
+            allPersonas,
             setMinimalMode(enabled) {
                 Services.Settings.setPluginMinimalModeEnabled(id, enabled)
             },
