+++ packages/mask/src/extension/popups/render.tsx
@@ -4,22 +4,9 @@ import { createNormalReactRoot } from '../../utils/index.js'
 import { createPluginHost, createSharedContext } from '../../../shared/plugin-infra/host.js'
 import Services from '#services'
 import Popups from './UI.js'
-import { NextSharedUIContext, RestPartOfPluginUIContextShared } from '../../utils/plugin-context-shared-ui.js'
-import { MaskMessages, createSubscriptionFromAsync } from '@masknet/shared-base'
-import { __setUIContext__ } from '@masknet/plugin-infra/dom'
+import { allPersonas } from '../../../shared-ui/initUIContext.js'
 
 if (location.hash === '') location.assign('#/personas')
-__setUIContext__({
-    allPersonas: NextSharedUIContext.allPersonas,
-    currentPersona: NextSharedUIContext.currentPersonaIdentifier,
-    queryPersonaAvatar: Services.Identity.getPersonaAvatar,
-    querySocialIdentity: Services.Identity.querySocialIdentity,
-    fetchJSON: Services.Helper.fetchJSON,
-    queryPersonaByProfile: Services.Identity.queryPersonaByProfile,
-    openDashboard: Services.Helper.openDashboard,
-    openPopupWindow: Services.Helper.openPopupWindow,
-    signWithPersona: (a, b, c, d) => Services.Identity.signWithPersona(a, b, c, location.origin, d),
-})
 
 /**
  * Firefox will not help popup fixed width when user click browser action
@@ -47,27 +34,13 @@ startPluginHost()
 function startPluginHost() {
     // TODO: Should only load plugins when the page is plugin-aware.
 
-    const allPersonaSub = createSubscriptionFromAsync(
-        () => Services.Identity.queryOwnedPersonaInformation(true),
-        [],
-        (x) => {
-            const clearCurrentPersonaIdentifier = MaskMessages.events.currentPersonaIdentifier.on(x)
-            const clearOwnPersonaChanged = MaskMessages.events.ownPersonaChanged.on(x)
-
-            return () => {
-                clearCurrentPersonaIdentifier()
-                clearOwnPersonaChanged()
-            }
-        },
-    )
-
     startPluginDashboard(
         createPluginHost(
             undefined,
             (id, def, signal) => ({
                 ...createSharedContext(id, signal),
-                ...RestPartOfPluginUIContextShared,
-                allPersonas: allPersonaSub,
+                allPersonas,
+                hasPaymentPassword: Services.Wallet.hasPassword,
                 setMinimalMode(enabled) {
                     Services.Settings.setPluginMinimalModeEnabled(id, enabled)
                 },
