+++ packages/plugin-infra/src/site-adaptor/context.ts
@@ -3,11 +3,10 @@
 
 import type { Subscription } from 'use-subscription'
 import type { IdentityResolved } from '../types.js'
-import { __setUIContext__, type __UIContext__ } from '../dom/context.js'
 import type { NextIDPlatform, PersonaIdentifier, PostIdentifier } from '@masknet/shared-base'
 
 export * from '../dom/context.js'
-export interface __SiteAdaptorContext__ extends __UIContext__ {
+export interface __SiteAdaptorContext__ {
     lastRecognizedProfile: Subscription<IdentityResolved | undefined>
     currentVisitingProfile: Subscription<IdentityResolved | undefined>
     currentNextIDPlatform: NextIDPlatform | undefined
@@ -24,7 +23,6 @@ export let getPostURL: __SiteAdaptorContext__['getPostURL']
 export let share: __SiteAdaptorContext__['share']
 export let getUserIdentity: __SiteAdaptorContext__['getUserIdentity']
 export function __setSiteAdaptorContext__(value: __SiteAdaptorContext__) {
-    __setUIContext__(value)
     ;({
         lastRecognizedProfile,
         currentVisitingProfile,
