+++ packages/shared/src/hooks/useNextIDVerify.ts
@@ -11,11 +11,11 @@ import {
 } from '@masknet/shared-base'
 import { NextIDProof } from '@masknet/web3-providers'
 import { useSiteAdaptorContext } from '@masknet/plugin-infra/content-script'
-import { currentNextIDPlatform } from '@masknet/plugin-infra/content-script/context'
+import { currentNextIDPlatform, signWithPersona } from '@masknet/plugin-infra/content-script/context'
 
 export function useNextIDVerify() {
     const verifyPostCollectTimer = useRef<ReturnType<typeof setTimeout> | null>(null)
-    const { getPostIdFromNewPostToast, postMessage, signWithPersona } = useSiteAdaptorContext()
+    const { getPostIdFromNewPostToast, postMessage } = useSiteAdaptorContext()
 
     return useAsyncFn(
         async (persona?: PersonaInformation, username?: string, verifiedCallback?: () => void | Promise<void>) => {
